requirements.txt
antlr4-python3-runtime==4.9.3
autocorrect==2.6.1
blis==0.7.7
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Command==0.1.0
cycler==0.11.0
cymem==2.0.6
cytoolz==0.11.2
emoji==1.7.0
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.2.0/es_core_news_lg-3.2.0-py3-none-any.whl
es-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/es_dep_news_trf-3.2.0/es_dep_news_trf-3.2.0-py3-none-any.whl
filelock==3.6.0
fonttools==4.33.3
gensim==4.1.2
huggingface-hub==0.5.0
hydra-core==1.2.0
idna==3.3
imbalanced-learn==0.9.1
imblearn==0.0
inexactsearch==1.0.2
jellyfish==0.9.0
jieba==0.42.1
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.2
murmurhash==1.0.6
networkx==2.7.1
nltk==3.7
numpy==1.22.3
omegaconf==2.2.2
packaging==21.3
pandas==1.4.1
pathy==0.6.1
Pillow==9.1.1
preshed==3.0.6
protobuf==3.20.0
psutil==5.9.0
pydantic==1.8.2
pyparsing==3.0.7
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.1.1
scikit-plot==0.3.7
scipy==1.8.0
seaborn==0.11.2
silpa-common==0.3
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soundex==1.1.3
spacy==3.2.4
spacy-alignments==0.8.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-stanza==1.0.1
spacy-transformers==1.1.5
spellchecker==0.4
srsly==2.4.2
stanza==1.3.0
textacy==0.12.0
textblob==0.17.1
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.11.6
toolz==0.11.2
torch==1.11.0
tqdm==4.64.0
transformers==4.17.0
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.1
